section 5 Identity Server complete

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -43,13 +43,13 @@ public class AuctionsController : ControllerBase
         return auction;
     }
 
-    //[Authorize]
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
     {
         var auction = _mapper.Map<Auction>(auctionDto);
 
-        auction.Seller = "test"; //User.Identity.Name;
+        auction.Seller = User.Identity.Name;
 
         _repo.AddAuction(auction);
 
@@ -64,7 +64,7 @@ public class AuctionsController : ControllerBase
             new { auction.Id }, newAuction);
     }
 
-    //[Authorize]
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
     {
@@ -72,7 +72,7 @@ public class AuctionsController : ControllerBase
 
         if (auction == null) return NotFound();
 
-        // if (auction.Seller != User.Identity.Name) return Forbid();
+        if (auction.Seller != User.Identity.Name) return Forbid();
 
         auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
         auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -89,7 +89,7 @@ public class AuctionsController : ControllerBase
         return BadRequest("Problem saving changes");
     }
 
-    //[Authorize]
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
@@ -97,7 +97,7 @@ public class AuctionsController : ControllerBase
 
         if (auction == null) return NotFound();
 
-        //if (auction.Seller != User.Identity.Name) return Forbid();
+        if (auction.Seller != User.Identity.Name) return Forbid();
 
         _repo.RemoveAuction(auction);
